express/requirements.txt
django-crispy-forms
djangorestframework
django-tables2
django-filter
django-excel
pytz
coreapi
pyexcel-xls
pyexcel-xlsx
reportlab
pypinyin
celery
django-celery-results
flower
django-celery-beat
XlsxWriter
dicttoxml
PyPDF2
Wand
xmltodict
django-ckeditor